Feature/back/board

See merge request s05-blockchain/S05P21B102!26

+++ Backend/init/src/main/java/com/blockback/init/service/UserServiceImpl.java
@@ -6,6 +6,7 @@ import com.blockback.init.entity.User;
 import com.blockback.init.repository.UserRepository;
 import com.blockback.init.repository.UserRepositorySup;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.ClassPathResource;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestPart;
@@ -51,7 +52,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User createUser(UserRegisterPostReq userRegisterInfo, MultipartFile thumbnail) {
-        String BASE_PATH = System.getProperty("user.dir") + "/image/user/";
+        String BASE_PATH = System.getProperty("user.dir") + "/init/src/main/resources/image/user/";
+        System.out.println(BASE_PATH + "??");
+
         User user = new User();
         try {
             /*
@@ -62,7 +65,8 @@ public class UserServiceImpl implements UserService {
                 String filePath = BASE_PATH + "default.png";
                 user.setUserthumbnail(filePath);
             } else { // 프로필 사진 입력함
-                String filePath = BASE_PATH + userRegisterInfo.getUserEmail() + "-" + thumbnail.getOriginalFilename();
+                String filePath = BASE_PATH + userRegisterInfo.getUserEmail() + "_" + thumbnail.getOriginalFilename();
+                System.out.println(filePath);
                 File dest = new File(filePath);
                 thumbnail.transferTo(dest);
                 if (!dest.exists()) { // 파일 존재 x, 이 부분이 현재 null 값이 들어갈 수 있음, 추후 다시 해야됨
@@ -72,6 +76,7 @@ public class UserServiceImpl implements UserService {
                 }
             }
             user.setUserEmail(userRegisterInfo.getUserEmail());
+            user.setUseraccount(userRegisterInfo.getUseraccount());
             // 보안을 위해서 유저 패스워드 암호화 하여 디비에 저장.
             user.setPassword(passwordEncoder.encode(userRegisterInfo.getPassword()));
             user.setUsernickname(userRegisterInfo.getUsernickname());
@@ -89,20 +94,26 @@ public class UserServiceImpl implements UserService {
     @Override
     public User putUser(UserPutReq putinfo, Long userId, MultipartFile thumbnail) {
         User user = userRepository.findById(userId).get();
-        String BASE_PATH = System.getProperty("user.dir") + "image/user/";
+        String BASE_PATH = System.getProperty("user.dir") + "/init/src/main/resources/image/user/";
+        String pic_place = user.getUserthumbnail();
         try {
             /*
              * 프로필 이미지 저장 순번-파일명으로 저장
              */
             if(thumbnail == null) { // 프로필 사진 입력 안함
                 //프로필 이미지 저장 (기본 이미지)
-                File before = new File(user.getUserthumbnail());
-                before.delete();
+                if (!pic_place.substring(pic_place.length()-11, pic_place.length()).equals("default.png")){
+                    // 없을 수도 있잖아 오류가 안난다.
+                    File before = new File(user.getUserthumbnail());
+                    before.delete();
+                }
                 String filePath = BASE_PATH + "default.png";
                 user.setUserthumbnail(filePath);
             } else { // 프로필 사진 입력함
-                File before = new File(user.getUserthumbnail());
-                before.delete();
+                if (!pic_place.substring(pic_place.length()-11, pic_place.length()).equals("default.png")){
+                    File before = new File(user.getUserthumbnail());
+                    before.delete();
+                }
                 String filePath = BASE_PATH + putinfo.getUserEmail() + "-" + thumbnail.getOriginalFilename();
                 File dest = new File(filePath);
                 thumbnail.transferTo(dest);
